Add Communicator constructor to JavaScript
