+++ js/src/Ice/Communicator.js
@@ -1,17 +1,55 @@
 // Copyright (c) ZeroC, Inc.
 
+import { CompositeSliceLoader } from "./CompositeSliceLoader.js";
+import { InitializationData } from "./InitializationData.js";
+import { InitializationException } from "./LocalExceptions.js";
+import { ObjectPrx } from "./ObjectPrx.js";
+import { Promise } from "./Promise.js";
+import { Properties } from "./Properties.js";
+import { defaultSliceLoaderInstance } from "./DefaultSliceLoader.js";
 import { generateUUID } from "./UUID.js";
 import { identityToString } from "./IdentityToString.js";
-import { Promise } from "./Promise.js";
-import { ObjectPrx } from "./ObjectPrx.js";
 
 //
 // Ice.Communicator
 //
 export class Communicator {
-    constructor(initData) {
+    constructor(arg1, arg2) {
         this._isShutdown = false;
         this._shutdownPromise = new Promise();
+
+        let args = null;
+        let initData = null;
+
+        if (arg1 instanceof Array) {
+            args = arg1;
+        } else if (arg1 instanceof InitializationData) {
+            initData = arg1;
+        } else if (arg1 !== undefined && arg1 !== null) {
+            throw new InitializationException("invalid argument to initialize");
+        }
+
+        if (arg2 !== undefined && arg2 !== null) {
+            if (arg2 instanceof InitializationData && initData === null) {
+                initData = arg2;
+            } else {
+                throw new InitializationException("invalid argument to initialize");
+            }
+        }
+
+        if (initData === null) {
+            initData = new InitializationData();
+        } else {
+            initData = initData.clone();
+        }
+        initData.properties = new Properties(args, initData.properties);
+
+        if (initData.sliceLoader === null || initData.sliceLoader === undefined) {
+            initData.sliceLoader = defaultSliceLoaderInstance;
+        } else {
+            initData.sliceLoader = new CompositeSliceLoader([initData.sliceLoader, defaultSliceLoaderInstance]);
+        }
+
         this._instance = this.createInstance(initData);
         this._instance.finishSetup(this);
     }
