+++ js/test/Ice/properties/Client.ts
@@ -116,7 +116,7 @@ export class Client extends TestHelper {
         }
 
         {
-            await using communicator = Ice.initialize(args);
+            await using communicator = new Ice.Communicator(args);
             const properties = communicator.getProperties();
 
             out.write("testing that creating an object adapter with unknown properties throws an exception...");
